backend/requirements.txt
langchain>=0.2.9
langchain
langchain-community
langchain-ollama
langchain-chroma
langchain-text-splitters
chromadb
fastapi
uvicorn
pypdf
pdfplumber
pytesseract
pillow
python-multipart
PyPDF2
pytesseract
pdf2image
python-docx
python-pptx
python-magic
fastapi-utils